+++ middleware/bodyParser/index.js
@@ -1,0 +1,4 @@
+const koaBodyparser = require('koa-bodyparser');
+module.exports = (options) => {
+    return koaBodyparser(options);
+};
