+++ middleware/cors/index.js
@@ -1,0 +1,4 @@
+const koaCors = require('koa-cors');
+module.exports = (options) => {
+    return koaCors(options);
+};
