+++ middleware/formParser/index.js
@@ -1,0 +1,4 @@
+const koaFormidable = require('koa2-formidable');
+module.exports = (options) => {
+    return koaFormidable(options);
+};
