+++ middleware/index.js
@@ -1,4 +1,8 @@
 module.exports = {
-    errors: require('./errors'),
-    validaor: require('./validator')
+    bodyParser: require('./bodyParser'),
+    cors: require('./cors'),
+    formParser: require('./formParser'),
+    requestHandler: require('./requestHandler'),
+    ui: require('./ui'),
+    validator: require('./validator')
 };
